Add data on the build and deployment workflows
This adds two pieces of metadata to the build and deployment process
that will exist in the deployed application as static files:
.deployment-data.json and .build-data.json. Perhaps having both files is
a little unnecessary; however, there is the possibility that at some
point their existence will be useful for troubleshooting pipeline
problems, validating that the deployed application is properly
up-to-date, and checking the version of the app deployed.

This _is not_ added as metadata to every page. The complexity for that
is larger and unnecessary if we can access a static file (that we could
stop including or limit access to trivially in the future).

With the current implementation it is true that the SHA and REF in both
files should match and that the timestamp should differ.