Add the basic elements for creating a generic data table base on Vuetify
as requete on the ticket on comment https://ccpo.atlassian.net/browse/AT-6503?focusedCommentId=319289
I am creating an generic table object that recives as props Data and sorting
Set up the data type to match the structure of the defoult V-data-table object on vuetify.
Crea a simple tamplate that recive this params to create the table
And add an style file for customize the styling and get it to look as requested.

Status: this changes have not yet been tested and it constitude my firts aproach to creating this feacture.

 AT-6503