Implementing and testing the generic table commponent
I using the step 4 page to test the looks and feel of the component ATATTable.

status: It seems like the defaulth vuetify dataTable element that i extending would not be able to replicate the look on the markups since it have various item per column so may be it will requeire something done more custom to that use case.

AT-6503