Basic Styles and cards added.
After meeting with Tony at 4:35 pm yesterday, he got me the go-ahead with my second idea. I added the table's external card on the step 4 view, title, and edit button and did the basic styling.

- Set all font sizes using the provided styles and Vuetify default styles.
- Set the shadow of the card to have the same parameter as in the markup.
- Set padding and margins to make the table as similar as possible as the markup.
- Resolve various style bugs and conflicts with Vuetify.

Status: at this point on the step 4 view, you can see a demo that looks pretty similar to the markup; And Idea number one, the generic table, is still available for future development but not in use since it is out of the scope of the ticket. The next step is to do the wave test, IE, test, clean up the code, and approve my style.

feature/AT-6503