Additional Styles bug fixes
Now the Card structure is part of the table
Alignment on the table is no longer vertically center to comply with the markup.
Fixing a bug with the shadows.
Test in IE looks good.
Test on Wave pass.

Status: All the looks and feels look correct; now I am back to work on the data structure, passing the title or application id, and after that, I might be able to start working on the jest coverage.

feature/AT-6503