Enable Dependabot
This enables dependabot for the base project, the CDK project
separately, and for GitHub Actions. It also ensures that both teams get
notified on dependency update PRs for GitHub Actions and for the CDK by
specifying both teams as reviwers.

Ideally, we would specify the Platform team as assignees for the CDK and
GitHub Actions PRs; however, GitHub does not currently permit assigning
a PR to a particular team (only individuals).