Create Cognito User Pool Client
This creates the client for authenticating to the user pool based on the
attributes defined during the inital proof of concept work. The values
here may need to be adjusted to our specific use case.

At the same time, the versions of the CDK and other dependencies in use
have been increased and the package-lock.json has been forced into
version control to comply with general best-practices.