Fix TypeScript compilation errors
It seems like these two issues made TypeScript somewhat grumpy recently.
Their trivial issues. For some reason, TypeScript's compiler is happier
to see us throw away the result of calling the stack's constructor than
to save it to a variable. And technically the type of `string &&
boolean` expression is `string | boolean` so we need to coerce the
string to a boolean and `!!str` does just fine at that.

Both of these issues prevented a successful deployment and these are the
exact changes that were deployed. I would prefer to see this merged and
resolve other issues in a followup PR.