Ensure lockfileVersion stays consistent
This ensures that lockfileVersion == 1 for all package-lock.json files
in the repository. This will help prevent "battles" over the format and
version of the file from different versions of npm and catch them during
the automated checks. This check has been in place on the atat-web-api
repo for awhile now.