Fixup proxy-related bugs
This ensures that we can handle routes with more than one level of
slashes as well as routes with URI-encoded spaces. Of course, this
assumes that we don't have files in S3 with URI-encoded names. Hopefully
we can assume that's the case going forward.

This also makes a change to the USAGovBanner to ensure that the two
icons work correctly with the change to `publicPath`. This uses the same
logic that is used for the flag.