Remove problematic regular expression
SCA suggested that the previously-used regular expression was vulnerable
to ReDoS attacks and because the "fixed" regex was either too
complicated or still not sufficient, we instead just multiple passes
over the string to get the results we need. In a lot of ways, this
actually feels a lot cleaner. In the worst cases, it may still be more
slow, but at least it's not exponential.