Fix building configuration URL
Because we previously returned if window.location.hostname was
"localhost", the ternary in this statement provided no value (since the
condition must always evaluate to false). We can instead operate
assuming that we are able to leverage `window.location` correctly.

This builds the configuration URL based on the origin, Vue.js
`publicPath` variable (exposed as `process.env.BASE_URL`), and
statically configured well-known value for the configuration endpoint
(`"configuration"`).

This also resolves a minor bug in the previous code that prevented the
configuration from being loaded if the user omitted the trailing `/` at
the end of the URL for the site. This also ensures that the
configuration can be loaded in the event that a user navigates somewhere
other than `/` first.