Properly retrieve IdP names
Unfortunately, there's a bug in the CDK where it tries to Fn::Sub over
an SSM resolve. The Fn::Sub happens before the resolve though which
means you get some wonky results. Instead, we need to just create a
CfnParameter resource and pull the value(s) from there.

This also adds an output for the bucket name to make deploying somewhat
easier.