nav updates
Set Acquisition Package child 'Project Overview' path to the same as parent path...   (This is necessary so that the default/first child view is display when parent route is displayed)
Updated ATATTextField default rules declaration to ()=>[]
updated ATATStepper with a helper method to get the route name when the route has children. "getRouteName" returns the name of the first child when the parent route is selected in the Stepper nav.