updates to stepper mapping and route resolver
I updated the stepper mapping for navigation so that nameless parent routes are not mapped
I updated the route resolver functionality is performed in a component called Resolver.  This should allow any data persistence that occurs in 'SaveOnLeave' to be finalized.  The Resolver component is never actually rendered because the Resolver component routes the UI to the correct stepper by calling the route resolver function BeforeRouteEnter