Fix coverage reporting for Vue with Istanbul
It seems that Istanbul, by default, is not configured to collect
coverage for `.vue` files. This adds that configuration. It also only
adds the plugin when `NODE_ENV` is `test` (should happen by default with
`npm run test`).

A custom name is required because it seems that `@vue/app` already
includes the istanbul plugin but misconfigured.