Add JWCC component unit tests
Add initial start to unit test for JWCC. This component relies
on various calls to different tables to get aggregate data for
task orders and funding. Some of the data modeling is chaning,
so might be better to do after those changes.

There will be a good amount of mocking for this test from the
gathered data, and it is best to mock the bigger functions
that return a majority of the data.

Update BarChart component to use async/await for mount to prevent
unit test from failing.