Fix JWCC Dashboard charts
The Costs table was still showing the `Service/Agency`
column causing requests to show undeified for `Agency` in the
sandbox instance.

- Refresh `Costs` table to ensure `Agency` column is avaialble`

The 'getEntityTotals' was returning undefiend for agency keys, which
causes the 'fundsSpentByAgency' to return '.name undefined' error.
This is due to agency no longer being a choice field, but a reference
to a record in the Agency table, and required requests to the agency
table use 'agency.title' not 'agency' to get only the title of the
refrenced agency record.

- Refactor requests to the Costs table to use 'agency.title' to
get only the title and not the ref data

Update model for CostsDTO

Ticket: AT-7965